python/requirements.txt
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
whisper==1.1.10
python-docx==0.8.11